fix confusing naming
